# swagger.json
{"openapi":"3.0.0","info":{"title":"API Superhéroes y Mascotas","version":"1.0.0","description":"Documentación de la API para superhéroes, mascotas y adopciones"},"servers":[{"url":"https://heroesymascotas-production-aba1.up.railway.app"}],"tags":[{"name":"Héroes","description":"Operaciones sobre superhéroes"},{"name":"Mascotas","description":"Operaciones sobre mascotas"},{"name":"Adopciones","description":"Operaciones sobre adopciones"},{"name":"MascotaVirtual","description":"Operaciones sobre mascotas virtuales"}],"components":{"securitySchemes":{"heroIdAuth":{"type":"apiKey","in":"header","name":"x-hero-id","description":"ID numérico del héroe autenticado"}}},"security":[{"heroIdAuth":[]}],"paths":{"/api/adopt/{petId}":{"post":{"tags":["Adopciones"],"summary":"Un héroe adopta una mascota (requiere autenticación)","parameters":[{"in":"header","name":"x-hero-id","required":true,"schema":{"type":"integer"},"description":"ID del héroe autenticado"},{"in":"path","name":"petId","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Mensaje de éxito"},"400":{"description":"Error de validación"}}}},"/api/adoptions":{"get":{"tags":["Adopciones"],"summary":"Obtiene todas las adopciones del héroe autenticado","parameters":[{"in":"header","name":"x-hero-id","required":true,"schema":{"type":"integer"},"description":"ID del héroe autenticado"}],"responses":{"200":{"description":"Lista de adopciones del héroe"},"401":{"description":"No autorizado"}}}},"/api/adoptions/{petId}":{"get":{"tags":["Adopciones"],"summary":"Obtiene una mascota por su ID","parameters":[{"in":"header","name":"x-hero-id","required":true,"schema":{"type":"integer"},"description":"ID del héroe autenticado"},{"in":"path","name":"petId","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Mascota encontrada"},"403":{"description":"No tienes permiso para ver esta mascota"},"404":{"description":"Mascota no encontrada"}}}},"/api/auth/login":{"post":{"tags":["Auth"],"summary":"Iniciar sesión como superhéroe","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"Token JWT","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"}}}}}},"400":{"description":"Error de validación"}}}},"/api/auth/register":{"post":{"tags":["Auth"],"summary":"Registrar un nuevo superhéroe","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"},"alias":{"type":"string"},"powers":{"type":"array","items":{"type":"string"}}}}}}},"responses":{"200":{"description":"Usuario registrado"},"400":{"description":"Error de validación"}}}},"/api/heroes":{"get":{"tags":["Héroes"],"summary":"Obtiene todos los héroes","responses":{"200":{"description":"Lista de héroes"}}},"post":{"tags":["Héroes"],"summary":"Agrega un héroe","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"alias":{"type":"string"},"city":{"type":"string"},"team":{"type":"string"}}}}}},"responses":{"201":{"description":"Héroe agregado"},"400":{"description":"Error de validación"}}}},"/api/heroes/{id}":{"put":{"tags":["Héroes"],"summary":"Actualiza un héroe","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"alias":{"type":"string"},"city":{"type":"string"},"team":{"type":"string"}}}}}},"responses":{"200":{"description":"Héroe actualizado"},"404":{"description":"Héroe no encontrado"}}},"delete":{"tags":["Héroes"],"summary":"Elimina un héroe","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Héroe eliminado"},"404":{"description":"Héroe no encontrado"}}}},"/api/pets":{"get":{"tags":["Mascotas"],"summary":"Obtiene todas las mascotas","responses":{"200":{"description":"Lista de mascotas"}}},"post":{"tags":["Mascotas"],"summary":"Agrega una mascota","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string"},"power":{"type":"string"}}}}}},"responses":{"201":{"description":"Mascota agregada"},"400":{"description":"Error de validación"}}}},"/api/pets/{id}":{"put":{"tags":["Mascotas"],"summary":"Actualiza una mascota","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string"},"power":{"type":"string"}}}}}},"responses":{"200":{"description":"Mascota actualizada"},"404":{"description":"Mascota no encontrada"}}},"delete":{"tags":["Mascotas"],"summary":"Elimina una mascota","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Mascota eliminada"},"404":{"description":"Mascota no encontrada"}}}},"/api/virtual/virtual-pets":{"get":{"tags":["MascotaVirtual"],"summary":"Lista todas las mascotas virtuales","responses":{"200":{"description":"Lista de mascotas virtuales"}}},"post":{"tags":["MascotaVirtual"],"summary":"Crea una mascota virtual desde una mascota adoptada","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"petId":{"type":"integer"}}}}}},"responses":{"200":{"description":"Mascota virtual creada"},"400":{"description":"Error de validación"}}}},"/api/virtual/virtual-pets/{id}/estado":{"get":{"tags":["MascotaVirtual"],"summary":"Ver estado de una mascota virtual","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Estado de la mascota virtual"},"404":{"description":"Mascota virtual no encontrada"}}}},"/api/virtual/virtual-pets/{id}/actividad":{"post":{"tags":["MascotaVirtual"],"summary":"Realizar actividad (jugar, pasear) con la mascota virtual","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"activity":{"type":"string","example":"jugar"}}}}}},"responses":{"200":{"description":"Mascota virtual actualizada"},"400":{"description":"Error de validación"}}},"get":{"tags":["MascotaVirtual"],"summary":"Obtener historial de actividad de la mascota virtual","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Historial de actividad"},"404":{"description":"Mascota virtual no encontrada"}}}},"/api/virtual/virtual-pets/{id}/comida":{"post":{"tags":["MascotaVirtual"],"summary":"Alimentar mascota virtual","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"comida":{"type":"string","example":"pizza"}}}}}},"responses":{"200":{"description":"Mascota virtual alimentada"},"400":{"description":"Error de validación"}}}},"/api/virtual/virtual-pets/{id}/ropa":{"post":{"tags":["MascotaVirtual"],"summary":"Personalizar ropa de la mascota virtual","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"ropa":{"type":"string","example":"sombrero"}}}}}},"responses":{"200":{"description":"Mascota virtual actualizada"},"400":{"description":"Error de validación"}}}},"/api/virtual/virtual-pets/{id}/curar":{"post":{"tags":["MascotaVirtual"],"summary":"Curar mascota virtual","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Mascota virtual curada"},"400":{"description":"Error de validación"}}}},"/api/virtual/virtual-pets/{id}/muerta":{"get":{"tags":["MascotaVirtual"],"summary":"Ver si la mascota virtual está muerta","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Estado de muerte de la mascota virtual"},"400":{"description":"Error de validación"}}}},"/api/virtual/virtual-pets/{id}/revivir":{"post":{"tags":["MascotaVirtual"],"summary":"Revivir una mascota virtual muerta","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Mascota virtual revivida"},"400":{"description":"Error de validación"}}}}}}